Swap bazil.org/fuse to github.com/seaweedfs/fuse.
The bazil.org FUSE module dropped macOS support in April 2020 in response to
the closing of the source for the osxfuse project. While I'm sympathetic to the
principle of that decision, it meant I had to pin to an older commit to keep
things working.

Changes in the macFUSE 4.x series broke the older code, and my quick and dirty
attempts to patch it failed. This fork still maintains macFUSE support, at
least for now, and appears to be both drop-in compatible with the bazil.org
version, and also still supported. For now.