ffuse: release children after flushing files
Previously, the filesystem would cache file values indefinitely.  This is nice
for performance, but acts as a memory leak, since in the limit you'll page in
the entire filesystem.

Now, after successfully flushing a file on release, we'll drop any valid
children. This might be a little TOO aggressive, but to do something in between
will require more state tracking.