ffuse: add ffs.dataHash and ffs.dataHash.b64 magic xattrs
These xattrs compute a sha3/256 digest of the data of the file, not including
the node metadata. The resulting digest is computed from the content addresses
of the data blocks, so it is not a digest of the actual content, but it will
change if any bit of the content is modified. This formulation allows the hash
to be computed without reading the data blocks directly.