ffuse: port to github.com/hanwen/go-fuse
As a preparatory step toward supporting fuse-t, switch away from the
seaweedfs/fuse library to hanwen/go-fuse. The latter has much better support
for inode tracking and invalidation, and I like the API for filesystem
integration better as well.

As always, macOS support is imperfect, but it seems to handle all the cases
that matter for ffuse.