ffuse: allow hard linking arbitrary files to the tree
Add a new write-only magic xattr prefix, "ffs.link.<name>".

Setting "ffs.link.<name>" to <storage-key> on a directory D causes the
specified <name> to be added (or replaced) as a child of D, with the file whose
storage key is given by the xattr value.

Following the semantics of xattrs, the XATTR_CREATE flag requires that the name
not already exist, and XATTR_REPLACE requires that the name does already exist.
In any case, the operation fails if the specified storage key is not found.