ffuse: invalidate changed entries in ffs.link updates
This is slightly subtle, because the notification cannot happen in the midst of
the fs handler without a deadlock. So here we do it in a separate goroutine,
and rely on the fact that even if it doesn't happen right away, it will happen
"soon" and that is sufficient to make sure callers notice the change.